00020 - auth handler tests refactor
