+++ handlers/auth.handler.test.js
@@ -1,71 +1,10 @@
 const request = require('supertest')
 const AppHandler = require('./app.handler')
+const { userRepositoryMysql } = require('../test-helpers/user.repository.mock')
+const { securityHandler } = require('../test-helpers/security.handler.mock')
+const { bookRepository } = require('../test-helpers/book.repository.mock')
+const { readedBookRepository } = require('../test-helpers/book.repository.mock')
 
-getUserByEmail = jest.fn( email => {
-    if (email[0] === 'wrong_email@hotmail.com') {
-        return [[]]
-    }
-    return [[{name: 'username', password: '$2a$10$cY3QThX9jhUO4kFbvSfnWuWLHnWRkD8uzORVhkQjB7FbKtQujnac2'}]]
-})
-validatePassword = jest.fn( params => {
-    if (params === 'PasswordPruebaTesting123?*') {
-        return true;
-    }
-    return false
-})
-generateJsonWebToken = jest.fn( params => {
-    return 'jsonWebToken'
-})
-googleVerify = jest.fn( params => {
-    if (params === 'mytoken') {
-        return {
-            name: 'name',
-            email: 'email@hotmail.com',
-            given_name: 'given_name',
-            family_name: 'family_name',
-            jti: 'jti'
-        }
-    }
-    if (params === 'mytokennewuser') {
-        return {
-            name: 'name',
-            email: 'wrong_email@hotmail.com',
-            given_name: 'given_name',
-            family_name: 'family_name',
-            jti: 'jti'
-        }
-    }
-    throw new Error('Wrong token')
-})
-encryptPassword = jest.fn( params => {
-    return 'encryptedPassword'
-})
-createUser = jest.fn( params => {
-    return [
-        { insertId: 1 }
-    ]
-})
-validateJsonWebToken = jest.fn( (request, response, next) => {
-    next()
-})
-
-const userRepositoryMysql = {
-    getUserByEmail,
-    createUser
-}
-const securityHandler = {
-    validatePassword,
-    generateJsonWebToken,
-    googleVerify,
-    encryptPassword,
-    validateJsonWebToken
-}
-const bookRepository = {
-
-}
-const readedBookRepository = {
-
-}
 const appHandler = new AppHandler({userRepositoryMysql, securityHandler, bookRepository, readedBookRepository})
 const app = appHandler.createApp()
 describe('Login tests', () => {
