+++ test-helpers/book.repository.mock.js
@@ -1,0 +1,7 @@
+const bookRepository = {
+
+}
+
+module.exports = {
+    bookRepository
+};
