+++ test-helpers/readed-book.repository.mock.js
@@ -1,0 +1,7 @@
+const readedBookRepository = {
+
+}
+
+module.exports = {
+    readedBookRepository
+};
